requirements.txt
PyExifTool==0.5.2
pymediainfo==5.1.0
PyYAML==6.0
